OGM-910 Move the index creation from the dialect to the SchemaDefiner
