+++ core/src/main/java/org/hibernate/ogm/dialect/spi/GridDialect.java
@@ -8,7 +8,6 @@ package org.hibernate.ogm.dialect.spi;
 
 import org.hibernate.LockMode;
 import org.hibernate.dialect.lock.LockingStrategy;
-import org.hibernate.ogm.index.OgmIndexSpec;
 import org.hibernate.ogm.model.key.spi.AssociationKey;
 import org.hibernate.ogm.model.key.spi.AssociationKeyMetadata;
 import org.hibernate.ogm.model.key.spi.EntityKey;
@@ -188,12 +187,4 @@ public interface GridDialect extends Service {
 	 */
 	DuplicateInsertPreventionStrategy getDuplicateInsertPreventionStrategy(EntityKeyMetadata entityKeyMetadata);
 
-	/**
-	 * If IndexSpec are to be found in the {@link org.hibernate.ogm.datastore.spi.SchemaDefiner.SchemaDefinitionContext}
-	 * once validated, the {@link org.hibernate.ogm.datastore.spi.BaseSchemaDefiner} will trigger this to create the
-	 * associated Index in the datastore.
-	 *
-	 * @param indexSpec the datastore index specifications
-	 */
-	void createIndex(OgmIndexSpec indexSpec);
 }
