--- core/src/main/java/org/hibernate/ogm/index/OgmIndexSpec.java
@@ -1,16 +1,0 @@
-/*
- * Hibernate OGM, Domain model persistence for NoSQL datastores
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.ogm.index;
-
-/**
- * Index Specification Interface
- *
- * @author Francois Le Droff
- */
-
-public interface OgmIndexSpec {
-}
